Also rename context js files
