+++ webapp/src/routes/link-external-account/link-external-account.js
@@ -2,7 +2,8 @@ import React, { useContext } from "react";
 import Form from "react-jsonschema-form";
 import Button from "react-bootstrap/Button";
 
-import ProvidersContext from "context/LinkedAccountContext";
+import ProvidersContext from "context/linked-account-context";
+
 
 const LinkExternalAccount = () => {
